fix: use cached metadata for empty page previews
- Added a `preview` getter/setter to the `Item` class in `app-schema.ts` and `yjs-schema.ts` to cache page previews directly within the main project document.
- Extracted the `extractPagePreview` function into a reusable `pagePreview.ts` library module.
- Updated `OutlinerBase.svelte` to implement a debounced Yjs observer that automatically computes and saves the page preview to the document metadata upon user edits.
- Modified `PageListItem.svelte` to read the cached `page.preview` metadata first (bypassing the need for subdocument loading) and introduced an aesthetic empty state placeholder ("Empty page - Click to start writing") for pages with no content, resolving the issue where text-only pages appeared as "No content" in the project list view.
